# Documentation/api/openapi/weatherwear-api.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WeatherWear API",
    "description": "REST API for user authentication, weather lookup, AI clothing recommendations, recommendation history, feedback, and an AI style chat assistant.",
    "version": "1.0.0",
    "contact": {
      "name": "Dubrovin Mikita",
      "email": "dubrovin.nikita@student.ehu.lt"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8090/api",
      "description": "Local development server"
    },
    {
      "url": "https://wwtodipl-production.up.railway.app/api",
      "description": "Production server configured by application-prod.yml"
    }
  ],
...
  "paths": {
    "/auth/register": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "operationId": "registerUser",
        "summary": "Register a new user",
        "description": "Creates a user account with role USER and returns a JWT token for immediate use.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              },
              "examples": {
                "default": {
                  "summary": "Register with email and password",
                  "value": {
                    "email": "alex@example.com",
                    "password": "securePass123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                },
                "examples": {
                  "created": {
                    "value": {
                      "token": "eyJhbGciOiJIUzI1NiJ9.example.signature",
                      "email": "alex@example.com",
                      "role": "USER"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "409": {
            "$ref": "#/components/responses/ConflictError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "operationId": "loginUser",
        "summary": "Log in",
        "description": "Authenticates an existing user and returns a JWT token.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              },
              "examples": {
                "default": {
                  "value": {
                    "email": "alex@example.com",
                    "password": "securePass123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                },
                "examples": {
                  "success": {
                    "value": {
                      "token": "eyJhbGciOiJIUzI1NiJ9.example.signature",
                      "email": "alex@example.com",
                      "role": "USER"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/users/me": {
      "delete": {
        "tags": [
          "Account"
        ],
        "operationId": "deleteCurrentUser",
        "summary": "Delete current user account",
        "description": "Deletes the authenticated user account. Database foreign keys cascade the user's preferences, recommendation history, feedback, chat sessions, and chat messages.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Account and associated user-owned data deleted successfully. No response body is returned."
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT returned by /auth/register or /auth/login. Send it as Authorization: Bearer <token>."
      }
    },
    "responses": {
      "ValidationError": {
        "description": "Request validation failed, a required parameter is missing, or the JSON body is malformed.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "validation": {
                "value": {
                  "timestamp": "2026-04-30T12:45:00",
                  "status": 400,
                  "error": "Bad Request",
                  "message": "email: Email must be valid",
                  "path": "/api/auth/register"
                }
              }
            }
          }
        }
      },
      "UnauthorizedError": {
        "description": "Authentication is missing, invalid, or expired.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "unauthorized": {
                "value": {
                  "timestamp": "2026-04-30T12:45:00",
                  "status": 401,
                  "error": "Unauthorized",
                  "message": "Authentication is required",
                  "path": "/api/preferences"
                }
              }
            }
          }
        }
      },
...
      "ConflictError": {
        "description": "The request conflicts with existing data or database constraints.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "duplicateUser": {
                "value": {
                  "timestamp": "2026-04-30T12:45:00",
                  "status": 409,
                  "error": "Conflict",
                  "message": "User with email alex@example.com already exists",
                  "path": "/api/auth/register"
                }
              }
            }
          }
        }
      },
...
      "ServerError": {
        "description": "Unexpected server error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "serverError": {
                "value": {
                  "timestamp": "2026-04-30T12:45:00",
                  "status": 500,
                  "error": "Internal Server Error",
                  "message": "Unexpected server error",
                  "path": "/api/recommendations"
                }
              }
            }
          }
        }
      }
    },
    "schemas": {
      "RegisterRequest": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "Unique user email.",
            "example": "alex@example.com"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "format": "password",
            "description": "Plain password submitted by the client. The backend stores only a BCrypt hash.",
            "example": "securePass123"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "alex@example.com"
          },
          "password": {
            "type": "string",
            "format": "password",
            "example": "securePass123"
          }
        }
      },
      "AuthResponse": {
        "type": "object",
        "required": [
          "token",
          "email",
          "role"
        ],
        "properties": {
          "token": {
            "type": "string",
            "description": "JWT access token.",
            "example": "eyJhbGciOiJIUzI1NiJ9.example.signature"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "alex@example.com"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "timestamp",
          "status",
          "error",
          "message",
          "path"
        ],
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2026-04-30T12:45:00"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "example": 400
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          },
          "message": {
            "type": "string",
            "example": "email: Email must be valid"
          },
          "path": {
            "type": "string",
            "example": "/api/auth/register"
          }
        }
      },
      "Role": {
        "type": "string",
        "enum": [
          "USER",
          "ADMIN"
        ],
        "example": "USER"
      },
...
    }
  }
}